tests: use 64bit for ticks

+++ tests/unittests/tests-rtt_rtc/mock_rtt.c
@@ -87,7 +87,7 @@ static void _tick(void)
     }
 }
 
-void rtt_add_ticks(uint32_t ticks)
+void rtt_add_ticks(uint64_t ticks)
 {
     while (ticks--) {
         _tick();
